requirements.txt
pandas==2.2.2
reportlab==4.1.0
Pillow==10.3.0
segno==1.5.1